refactor 패키지 구조 변경
- infra 패키지를 adapter 패키지 안으로 이동
- 기존 adapter 패키지의 controller 부분을 client 패키지 생성 후 해당 패키지 안으로 이동
