+++ src/main/java/yapp/buddycon/app/gifticon/adapter/client/GifticonController.java
@@ -1,4 +1,4 @@
-package yapp.buddycon.app.gifticon.adapter;
+package yapp.buddycon.app.gifticon.adapter.client;
 
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
@@ -6,8 +6,8 @@ import org.springframework.data.domain.Page;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import yapp.buddycon.app.gifticon.adapter.request.SearchGifticonDTO;
-import yapp.buddycon.app.gifticon.adapter.response.GifticonResponseDTO;
+import yapp.buddycon.app.gifticon.adapter.client.request.SearchGifticonDTO;
+import yapp.buddycon.app.gifticon.adapter.client.response.GifticonResponseDTO;
 import yapp.buddycon.app.gifticon.application.port.in.GifticonUseCase;
 
 @RestController
