+++ src/main/java/yapp/buddycon/app/gifticon/adapter/client/request/SearchGifticonSortType.java
@@ -1,4 +1,4 @@
-package yapp.buddycon.app.gifticon.adapter.request;
+package yapp.buddycon.app.gifticon.adapter.client.request;
 
 public enum SearchGifticonSortType {
