+++ src/main/java/yapp/buddycon/app/gifticon/adapter/infra/GifticonRepository.java
@@ -1,12 +1,12 @@
-package yapp.buddycon.app.gifticon.infra;
+package yapp.buddycon.app.gifticon.adapter.infra;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Repository;
-import yapp.buddycon.app.gifticon.adapter.request.SearchGifticonDTO;
-import yapp.buddycon.app.gifticon.adapter.response.GifticonResponseDTO;
-import yapp.buddycon.app.gifticon.infra.jpa.GifticonJpaRepository;
+import yapp.buddycon.app.gifticon.adapter.client.request.SearchGifticonDTO;
+import yapp.buddycon.app.gifticon.adapter.client.response.GifticonResponseDTO;
+import yapp.buddycon.app.gifticon.adapter.infra.jpa.GifticonJpaRepository;
 import yapp.buddycon.app.gifticon.application.port.out.GifticonQueryPort;
 
 @RequiredArgsConstructor
