+++ src/main/java/yapp/buddycon/app/gifticon/adapter/infra/jpa/GifticonJpaRepositoryImpl.java
@@ -1,4 +1,4 @@
-package yapp.buddycon.app.gifticon.infra.jpa;
+package yapp.buddycon.app.gifticon.adapter.infra.jpa;
 
 import com.querydsl.core.types.OrderSpecifier;
 import com.querydsl.core.types.dsl.BooleanExpression;
@@ -12,12 +12,12 @@ import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort.Direction;
 import org.springframework.data.jpa.repository.support.QuerydslRepositorySupport;
-import yapp.buddycon.app.gifticon.adapter.request.SearchGifticonDTO;
-import yapp.buddycon.app.gifticon.adapter.response.GifticonResponseDTO;
+import yapp.buddycon.app.gifticon.adapter.client.request.SearchGifticonDTO;
+import yapp.buddycon.app.gifticon.adapter.client.response.GifticonResponseDTO;
+import yapp.buddycon.app.gifticon.adapter.client.response.QGifticonResponseDTO;
 import yapp.buddycon.app.gifticon.domain.Gifticon;
 import yapp.buddycon.app.gifticon.domain.QGifticon;
-import yapp.buddycon.app.gifticon.adapter.request.SearchGifticonSortType;
-import yapp.buddycon.app.gifticon.adapter.response.QGifticonResponseDTO;
+import yapp.buddycon.app.gifticon.adapter.client.request.SearchGifticonSortType;
 import yapp.buddycon.app.gifticon.domain.QStore;
 import yapp.buddycon.app.gifticon.domain.QStoreCategory;
