+++ src/main/java/yapp/buddycon/app/gifticon/application/port/in/GifticonUseCase.java
@@ -1,8 +1,8 @@
 package yapp.buddycon.app.gifticon.application.port.in;
 
 import org.springframework.data.domain.Page;
-import yapp.buddycon.app.gifticon.adapter.request.SearchGifticonDTO;
-import yapp.buddycon.app.gifticon.adapter.response.GifticonResponseDTO;
+import yapp.buddycon.app.gifticon.adapter.client.request.SearchGifticonDTO;
+import yapp.buddycon.app.gifticon.adapter.client.response.GifticonResponseDTO;
 
 public interface GifticonUseCase {
