+++ src/test/java/yapp/buddycon/app/gifticon/GifticonServiceTest.java
@@ -16,8 +16,8 @@ import org.springframework.data.domain.Page;
 import java.util.Arrays;
 import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.Pageable;
-import yapp.buddycon.app.gifticon.adapter.request.SearchGifticonDTO;
-import yapp.buddycon.app.gifticon.adapter.response.GifticonResponseDTO;
+import yapp.buddycon.app.gifticon.adapter.client.request.SearchGifticonDTO;
+import yapp.buddycon.app.gifticon.adapter.client.response.GifticonResponseDTO;
 import yapp.buddycon.app.gifticon.application.port.out.GifticonQueryPort;
 import yapp.buddycon.app.gifticon.application.service.GifticonService;
